+++ R/group-by.r
@@ -79,7 +79,7 @@ group_by_ <- function(.data, ..., .dots, add = FALSE) {
 #'   \item{groups}{Modified groups}
 #' @noRd
 group_by_prepare <- function(.data, ..., .dots, add = FALSE) {
-  new_groups <- lazyeval::all_dots(.dots, ..., env = parent.frame())
+  new_groups <- lazyeval::all_dots(.dots, ...)
 
   # If any calls, use mutate to add new columns, then group by those
   is_name <- vapply(new_groups, function(x) is.name(x$expr), logical(1))