+++ NEWS.md
@@ -203,9 +203,7 @@
   supports cumulative functions (min, max, mean, sum) and throws an error
   if you attempt to use recycled aggregates (which bq doesn't support).
 
-* new `verb()` distinct returns distinct (unique) rows of a tbl. Currently
-  implemented for `tbl_df()`, `grouped_df()`, `tbl_dt()`, `grouped_dt()`, and 
-  `tbl_sql()`
+* new verb `distinct()` returns distinct (unique) rows of a tbl (#97).
 
 * `select()` gives error message if you attempt to select invalid columns
   (#348)